move lsp code to go
