+++ init.lua
@@ -43,93 +43,6 @@ require("lazy").setup(
   {}
 )
 
--- lsp
-local opts = { noremap=true, silent=true }
---vim.keymap.set('n', '<space>e', vim.diagnostic.open_float, opts)
---vim.keymap.set('n', '[d', vim.diagnostic.goto_prev, opts)
---vim.keymap.set('n', ']d', vim.diagnostic.goto_next, opts)
---vim.keymap.set('n', '<space>q', vim.diagnostic.setloclist, opts)
-
--- Use an on_attach function to only map the following keys
--- after the language server attaches to the current buffer
-local on_attach = function(client, bufnr)
-  -- Enable completion triggered by <c-x><c-o>
-  vim.api.nvim_buf_set_option(bufnr, 'omnifunc', 'v:lua.vim.lsp.omnifunc')
-
-  -- Mappings.
-  -- See `:help vim.lsp.*` for documentation on any of the below functions
-  local bufopts = { noremap=true, silent=true, buffer=bufnr }
-  -- Will jump to interface files in OCaml
-  vim.keymap.set('n', 'gD', vim.lsp.buf.declaration, bufopts)
-  vim.keymap.set('n', 'gd', vim.lsp.buf.definition, bufopts)
-  vim.keymap.set('n', 'K', vim.lsp.buf.hover, bufopts)
-  vim.keymap.set('n', 'gi', vim.lsp.buf.implementation, bufopts)
-  vim.keymap.set('n', '<space>rn', vim.lsp.buf.rename, bufopts)
-  vim.keymap.set('n', '<space>ca', vim.lsp.buf.code_action, bufopts)
-  vim.keymap.set('n', '<space>e', vim.diagnostic.open_float, opts)
-  vim.keymap.set('n', 'gr', vim.lsp.buf.references, bufopts)
-  vim.keymap.set('n', '<space>f', function() vim.lsp.buf.format {async = true } end, bufopts)
-end
---
---require('lspconfig').dartls.setup {
---  --cmd = { "dart", "language-server", "--protocol=lsp", "--instrumentation-log-file=/usr/local/google/home/fujino/dart-lsp.log" },
---  cmd = {"dart", "language-server", "--protocol=lsp"},
---  on_attach = on_attach,
---  -- https://github.com/dart-lang/sdk/blob/main/pkg/analysis_server/tool/lsp_spec/README.md
---  init_options = {
---    -- When set to true, workspace folders will be ignored and analysis will be performed based on the open files, as if no workspace was open at all. This allows opening large folders without causing them to be completely analyzed. Defaults to false.
---    onlyAnalyzeProjectsWithOpenFiles = true,
---    -- When set to false, completion will not include symbols that are not already imported into the current file. Defaults to true, though the client must additionally support workspace/applyEdit for these completions to be included.
---    suggestFromUnimportedLibraries = true,
---    closingLabels = true,
---    outline = true,
---    flutterOutline = true,
---  },
---  settings = {
---    dart = {
---      showTodos = false,
---      -- Whether to include code snippets (such as class, stful, switch) in code completion. When unspecified, snippets will be included.
---      enableSnippets = false,
---    },
---  },
---}
---
---require('lspconfig').bashls.setup {
---  on_attach = on_attach,
---}
---
--- go install golang.org/x/tools/gopls@latest
-require('lspconfig').gopls.setup {
-  on_attach = on_attach,
-}
---
---require('lspconfig').ts_ls.setup {
---  on_attach = on_attach,
---}
---
----- gem install --user-install ruby-lsp
---require('lspconfig').ruby_lsp.setup {
---  on_attach = on_attach,
---}
---
---require('lspconfig').zls.setup {
---  on_attach = on_attach,
---}
---
---require('lspconfig').ocamllsp.setup {
---  on_attach = on_attach,
---  filetypes = { 'ocaml' },
---  root_dir = require('lspconfig.util').root_pattern('*.opam'),
---}
---
---require('lspconfig').clangd.setup {
---  on_attach = on_attach,
---}
---
---require('lspconfig').cmake.setup {
---  on_attach = on_attach,
---}
---
 ---- Debug Adapter
 ---- https://github.com/mfussenegger/nvim-dap/blob/master/doc/dap.txt
 --
@@ -179,12 +92,6 @@ require('lspconfig').gopls.setup {
 --    command = "if argc() == 1 && isdirectory(argv()[0]) | execute 'cd' argv()[0] | endif",
 --  }
 --)
---vim.cmd([[
---autocmd FileType python set shiftwidth=4 tabstop=4
---autocmd FileType cs set shiftwidth=4 tabstop=4
---autocmd FileType go set noexpandtab nosmarttab
---autocmd FileType markdown set linebreak
---]])
 
 ---- https://devforum.roblox.com/t/unnamed-anonymous-functions-give-ambiguous-syntax-errors-if-previous-line-contains-a-function-call/1450059
 f = (function ()
